feat: 환전 신청 목록 구현완료 (#84)

+++ src/main/java/com/ant/hurry/boundedContext/coin/dto/ExchangeRequest.java
@@ -10,5 +10,5 @@ public class ExchangeRequest {
     private BankType bankType;
     private String accountNumber;
     private String holderName;
-    private long money;
+    private int money;
 }
